refactor: UnbondConfirmationDialog -> ConfirmationDialog
