+++ src/components/Dashboard.js
@@ -7,7 +7,6 @@ import {
 	Table,
 	TableContainer,
 	TableHead,
-	Link,
 	TableBody,
 	Box,
 	Tooltip
